requirements.txt
cyhunspell==2.0.2
Werkzeug==2.0.3
Flask-HTTPAuth==4.2.0
flask-sqlalchemy==2.4.4
flask==2.1.1
gensim==4.1.2
gunicorn==20.1.0
nltk==3.7
numpy==1.22.2
pandas==1.2.3
psycopg2-binary==2.8.6
pyyaml==5.4.1
scipy==1.8.0
sentry-sdk[flask]==1.00
itsdangerous==2.0.1
prometheus-flask-exporter==0.20.1
